并发缓存

+++ src/main/java/top/alertcode/adelina/framework/service/impl/CommonService.java
@@ -25,6 +25,7 @@ import java.util.List;
 import java.util.Objects;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.locks.ReentrantLock;
+import java.util.stream.Collectors;
 
 /**
  * @author fuqiang // TODO: 2019/6/4 未考虑并发场景
@@ -47,6 +48,10 @@ public class CommonService extends BaseService implements BeanFactoryAware, IBas
     private ConcurrentHashMap<Serializable, String> mapLock = new ConcurrentHashMap();
 
 
+    private final static String QUERY = "query";
+    private final static String DELETE = "delete";
+
+
     @Override
     public void setBeanFactory(BeanFactory bFactory) throws BeansException {
         if (beanFactory == null && bFactory != null) {
@@ -66,14 +71,11 @@ public class CommonService extends BaseService implements BeanFactoryAware, IBas
             return JsonUtils.readValue(tableCacheDao.get(getHName(clazz), Objects.toString(id)), clazz);
         }
         boolean lock = false;
-        final String lockKey = getHName(clazz) + id;
+        final String lockKey = QUERY + getHName(clazz) + id;
         try {
             lock = mapLock.putIfAbsent(lockKey, "true") == null;
             if (lock) {
-                T t = (T) super.getById(id);
-                log.info("从数据库中取数据：线程={}", Thread.currentThread().getId());
-                tableCacheDao.add(getHName(clazz), getId(t), JsonUtils.writeValueAsString(t));
-                return t;
+                return getT(id, getHName(clazz));
             } else {
                 log.info("没拿到lock 降级：{}", Thread.currentThread().getId());
                 throw new ProjectException("当前请求忙，请稍后再试");
@@ -92,6 +94,7 @@ public class CommonService extends BaseService implements BeanFactoryAware, IBas
      * @param <T>
      * @return
      */
+    @Override
     public <T> T cacheGetById(Class<T> clazz, Serializable id, Model model) {
         if (model == Model.SegmentLock) {
             return cacheGetByIdSegmentLock(clazz, id);
@@ -110,13 +113,20 @@ public class CommonService extends BaseService implements BeanFactoryAware, IBas
                 log.info("从缓存中取数据：线程={}", Thread.currentThread().getId());
                 return JsonUtils.readValue(tableCacheDao.get(getHName(clazz), Objects.toString(id)), clazz);
             }
-            T t = (T) super.getById(id);
-            log.info("从数据库中取数据：线程={}", Thread.currentThread().getId());
-            tableCacheDao.add(getHName(clazz), getId(t), JsonUtils.writeValueAsString(t));
-            return t;
+            return getT(id, getHName(clazz));
         } finally {
             lock.unlock();
         }
+    }
+
+    private <T> T getT(Serializable id, String hName) {
+        T t = (T) super.getById(id);
+        if (Objects.isNull(t)) {
+            return null;
+        }
+        log.info("从数据库中取数据：线程={}", Thread.currentThread().getId());
+        tableCacheDao.add(hName, getId(t), JsonUtils.writeValueAsString(t));
+        return t;
     }
 
     @Override
@@ -128,25 +138,33 @@ public class CommonService extends BaseService implements BeanFactoryAware, IBas
 
 
     @Override
-    @Transactional
-    public boolean cacheDeleteById(Class clazz, Serializable id) {
-        tableCacheDao.delete(getHName(clazz), id.toString());
-        return super.removeById(id);
+    @Transactional(rollbackFor = Exception.class)
+    public synchronized boolean cacheDeleteById(Class clazz, Serializable id) {
+        if (tableCacheDao.exists(getHName(clazz), Objects.toString(id))) {
+            log.info("删除缓存，线程：{}", Thread.currentThread().getId());
+            tableCacheDao.delete(getHName(clazz), id.toString());
+            log.info("删除数据库，线程：{}", Thread.currentThread().getId());
+            return super.removeById(id);
+        }
+        return false;
     }
 
     @Override
-    @Transactional
-    public boolean cacheUpdateById(Object entity) {
-        boolean b = super.updateById(entity);
-        tableCacheDao.delete(getHName(entity.getClass()), getId(entity));
-        tableCacheDao.add(getHName(entity.getClass()), getId(entity), JsonUtils.writeValueAsString(entity));
-        return b;
+    @Transactional(rollbackFor = Exception.class)
+    public synchronized boolean cacheUpdateById(Object entity) {
+        if (tableCacheDao.exists(getHName(entity.getClass()), getId(entity))) {
+            boolean b = super.updateById(entity);
+            tableCacheDao.delete(getHName(entity.getClass()), getId(entity));
+            tableCacheDao.add(getHName(entity.getClass()), getId(entity), JsonUtils.writeValueAsString(entity));
+            return b;
+        }
+        return false;
     }
 
 
     @Override
-    @Transactional
-    public <T> boolean cacheSaveBatch(Collection<T> entityList) {
+    @Transactional(rollbackFor = Exception.class)
+    public synchronized <T> boolean cacheSaveBatch(Collection<T> entityList) {
         super.saveBatch(entityList);
         HashMap<String, String> map = new HashMap<>();
         if (CollectionUtils.isNotEmpty(entityList)) {
@@ -167,10 +185,11 @@ public class CommonService extends BaseService implements BeanFactoryAware, IBas
      */
 
     @Override
-    public <T> boolean cacheRemove(Wrapper<T> queryWrapper) {
+    public synchronized <T> boolean cacheRemove(Class<T> clazz, Wrapper<T> queryWrapper) {
         List list = super.list(queryWrapper);
-        if (CollectionUtils.isNotEmpty(list)) {
-            cacheDeleteByIds(list);
+        Collection collect = CollectionUtils.collect(list, this::getId);
+        if (CollectionUtils.isNotEmpty(collect)) {
+            cacheDeleteByIds(clazz, collect);
         }
         return true;
     }
@@ -181,14 +200,17 @@ public class CommonService extends BaseService implements BeanFactoryAware, IBas
      * @param idList 主键ID列表
      */
     @Override
-    public <T> boolean cacheDeleteByIds(Collection<T> idList) {
-        boolean b = super.removeByIds(idList);
+    public synchronized <T> boolean cacheDeleteByIds(Class<T> clazz, Collection<? extends Serializable> idList) {
         if (CollectionUtils.isNotEmpty(idList)) {
-            for (T t : idList) {
-                tableCacheDao.delete(getHName(t.getClass()), getId(t));
+            List<T> list = tableCacheDao.mutiGet(getHName(clazz),
+                    idList.stream().map(Objects::toString).collect(Collectors.toList()));
+            if (CollectionUtils.isNotEmpty(list)) {
+                tableCacheDao.delete(getHName(clazz),
+                        list.stream().map(this::getId).collect(Collectors.toSet()).toArray(new String[]{}));
+                return super.removeByIds(idList);
             }
         }
-        return b;
+        return false;
     }
 
 
@@ -199,7 +221,7 @@ public class CommonService extends BaseService implements BeanFactoryAware, IBas
      * @param updateWrapper 实体对象封装操作类 {@link com.baomidou.mybatisplus.core.conditions.update.UpdateWrapper}
      */
     @Override
-    public <T> void cacheUpdate(T entity, Wrapper<T> updateWrapper) {
+    public synchronized <T> void cacheUpdate(T entity, Wrapper<T> updateWrapper) {
         List<T> list = super.list(updateWrapper);
         HashMap<String, String> map = new HashMap<>();
         if (CollectionUtils.isNotEmpty(list)) {
@@ -213,7 +235,7 @@ public class CommonService extends BaseService implements BeanFactoryAware, IBas
      * 根据ID 批量更新
      */
     @Override
-    public <T> void cacheUpdateBatchById(Collection<T> entityList) {
+    public synchronized <T> void cacheUpdateBatchById(Collection<T> entityList) {
         HashMap<String, String> map = new HashMap<>();
         if (CollectionUtils.isNotEmpty(entityList)) {
             super.updateBatchById(entityList);
@@ -222,8 +244,8 @@ public class CommonService extends BaseService implements BeanFactoryAware, IBas
     }
 
     @Override
-    @Transactional
-    public <T> void cacheTbUpdateBatch(Collection<T> entityList, HashMap<String, String> map) {
+    @Transactional(rollbackFor = Exception.class)
+    public synchronized <T> void cacheTbUpdateBatch(Collection<T> entityList, HashMap<String, String> map) {
         for (T t : entityList) {
             map.put(getId(t), JsonUtils.writeValueAsString(entityList));
         }
